requirements.txt
mysql-connector-python
rich
matplotlib
